extract deleteSuspendedSale
